+++ ASM/c/dungeon_info.c
@@ -12,22 +12,23 @@
 int dungeon_count = 13;
 
 dungeon_entry_t dungeons[] = {
-    {  0, 0, 0, 0, 1, 0x0F, "Deku",      {-1, -1, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_DEKU_TREE },
-    {  1, 0, 0, 0, 1, 0x1F, "Dodongo",   {-1, -1, -1, -1}, { 0, -1, -1, -1}, OPT_HINT_AREA_DODONGOS_CAVERN },
-    {  2, 0, 0, 0, 1, 0x0F, "Jabu",      {-1, -1, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_JABU_JABUS_BELLY },
+    {  0, 0, 0, 0, 1, 0x0F, "Deku",       "Deku Tree",              {-1, -1, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_DEKU_TREE },
+    {  1, 0, 0, 0, 1, 0x1F, "Dodongo",    "Dodongo's Cavern",       {-1, -1, -1, -1}, { 0, -1, -1, -1}, OPT_HINT_AREA_DODONGOS_CAVERN },
+    {  2, 0, 0, 0, 1, 0x0F, "Jabu",       "Jabu Jabu's Belly",      {-1, -1, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_JABU_JABUS_BELLY },
 
-    {  3, 1, 1, 0, 1, 0x1F, "Forest",    {-1, -1, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_FOREST_TEMPLE },
-    {  4, 1, 1, 0, 1, 0x1F, "Fire",      {-1, -1, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_FIRE_TEMPLE },
-    {  5, 1, 1, 0, 1, 0x1F, "Water",     {-1, -1, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_WATER_TEMPLE },
-    {  7, 1, 1, 0, 1, 0x1F, "Shadow",    { 4,  6,  7, -1}, { 4,  5,  6,  7}, OPT_HINT_AREA_SHADOW_TEMPLE },
-    {  6, 1, 1, 0, 1, 0x1F, "Spirit",    {11, 14, 12, -1}, {13, 15, -1, -1}, OPT_HINT_AREA_SPIRIT_TEMPLE },
+    {  3, 1, 1, 0, 1, 0x1F, "Forest",     "Forest Temple",          {-1, -1, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_FOREST_TEMPLE },
+    {  4, 1, 1, 0, 1, 0x1F, "Fire",       "Fire Temple",            {-1, -1, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_FIRE_TEMPLE },
+    {  5, 1, 1, 0, 1, 0x1F, "Water",      "Water Temple",           {-1, -1, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_WATER_TEMPLE },
+    {  7, 1, 1, 0, 1, 0x1F, "Shadow",     "Shadow Temple",          { 4,  6,  7, -1}, { 4,  5,  6,  7}, OPT_HINT_AREA_SHADOW_TEMPLE },
+    {  6, 1, 1, 0, 1, 0x1F, "Spirit",     "Spirit Temple",          {11, 14, 12, -1}, {13, 15, -1, -1}, OPT_HINT_AREA_SPIRIT_TEMPLE },
 
-    {  8, 1, 0, 0, 1, 0x07, "BotW",      { 3, -1, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_BOTTOM_OF_THE_WELL },
-    {  9, 0, 0, 0, 1, 0x07, "Ice",       { 1,  2, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_ICE_CAVERN },
-    { 12, 1, 0, 1, 0, 0x00, "Hideout",   {-1, -1, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_THIEVES_HIDEOUT },
-    { 11, 1, 0, 0, 0, 0x00, "GTG",       { 8,  9, 10, -1}, { 8,  9, 10, -1}, OPT_HINT_AREA_GERUDO_TRAINING_GROUND },
-    { 13, 1, 1, 0, 0, 0x00, "Ganon",     {16, 17, 18, 21}, {18, 19, 20, -1}, OPT_HINT_AREA_INSIDE_GANONS_CASTLE },
-    { 16, 1, 0, 0, 0, 0x00, "Chest Game",{-1, -1, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_MARKET },
+    {  8, 1, 0, 0, 1, 0x07, "BotW",       "Bottom of the Well",     { 3, -1, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_BOTTOM_OF_THE_WELL },
+    {  9, 0, 0, 0, 1, 0x07, "Ice",        "Ice Cavern",             { 1,  2, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_ICE_CAVERN },
+    { 12, 1, 0, 1, 0, 0x00, "Hideout",    "Thieves' Hideout",       {-1, -1, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_THIEVES_HIDEOUT },
+    { 11, 1, 0, 0, 0, 0x00, "GTG",        "Gerudo Training Ground", { 8,  9, 10, -1}, { 8,  9, 10, -1}, OPT_HINT_AREA_GERUDO_TRAINING_GROUND },
+    { 10, 0, 0, 0, 0, 0x00, "Tower",      "Ganon's Tower",          {-1, -1, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_UNKNOWN },
+    { 13, 1, 1, 0, 0, 0x00, "Ganon",      "Ganon's Castle",         {16, 17, 18, 21}, {18, 19, 20, -1}, OPT_HINT_AREA_INSIDE_GANONS_CASTLE },
+    { 16, 1, 0, 0, 0, 0x00, "Chest Game", "Treasure Box Shop",      {-1, -1, -1, -1}, {-1, -1, -1, -1}, OPT_HINT_AREA_MARKET },
 };
 
 boss_entry_t bosses[] = {
@@ -148,6 +149,7 @@ int d_right_dungeon_idx(int i) {
     if (!CFG_DUNGEON_INFO_SILVER_RUPEES || !CFG_DUNGEON_IS_MQ[DODONGO_ID]) dungeon_idx++; // skip DC
     if (dungeon_idx >= 2) dungeon_idx++; // skip Jabu
     if (dungeon_idx >= 9 && (!CFG_DUNGEON_INFO_SILVER_RUPEES || CFG_DUNGEON_IS_MQ[ICE_ID])) dungeon_idx++; // skip Ice
+    if (dungeon_idx > 11) dungeon_idx++; // skip Tower
     return dungeon_idx;
 }
 
@@ -326,9 +328,9 @@ void draw_world_info(z64_disp_buf_t* db) {
                     continue;
                 }
                 gDPPipeSync(db->p++);
-                dungeon_entry_t dungeon = dungeons[i];
+                dungeon_entry_t dungeon = i == 12 ? dungeons[i + 1] : dungeons[i];
                 top += font_height + padding;
-                text_print_size(db, dungeon.name, left, top, font_width, font_height);
+                text_print_size(db, dungeon.short_name, left, top, font_width, font_height);
             }
 
             // Draw the list of dungeons interiors.
@@ -401,9 +403,9 @@ void draw_world_info(z64_disp_buf_t* db) {
                     continue;
                 }
                 gDPPipeSync(db->p++);
-                dungeon_entry_t dungeon = dungeons[i];
+                dungeon_entry_t dungeon = i == 12 ? dungeons[i + 1] : dungeons[i];
                 top += font_height + padding;
-                text_print_size(db, dungeon.name, left, top, font_width, font_height);
+                text_print_size(db, dungeon.short_name, left, top, font_width, font_height);
             }
             // List of bosses, located in CFG_BOSSES after the first list of 12 for the dpad left menu.
             for (uint8_t i = 0; i < rows - 1; i++) {
@@ -482,7 +484,7 @@ void draw_world_info(z64_disp_buf_t* db) {
                 }
                 dungeon_entry_t dungeon = dungeons[i];
                 top += font_height + padding;
-                text_print_size(db, dungeon.name, left, top, font_width, font_height);
+                text_print_size(db, dungeon.short_name, left, top, font_width, font_height);
             }
             // Draw the area each dungeon is located in.
             for (uint8_t i = 0; i < rows - 1; i++) {
@@ -699,12 +701,12 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
         // Pre completed dungeons are grayed and crossed out.
         for (int i = 0; i < rows; i++) {
             gDPPipeSync(db->p++);
-            dungeon_entry_t* d = &(dungeons[i]);
+            dungeon_entry_t* d = &(dungeons[i + (i > 11 ? 1 : 0)]); // skip Tower
             bool empty = CFG_DUNGEON_PRECOMPLETED[d->index];
             int top = start_top + ((icon_size + padding) * i) + 1;
             if (empty) {
                 gDPSetPrimColor(db->p++, 0, 0, 0xFF, 0xFF, 0xFF, 0x7F);
-                uint16_t sizeRectangle = text_print_size(db, d->name, left, top, font_width, font_height) - left;
+                uint16_t sizeRectangle = text_print_size(db, d->short_name, left, top, font_width, font_height) - left;
                 gDPSetPrimColor(db->p++, 0, 0, 0xFF, 0xFF, 0xFF, 0xBF);
                 gDPSetCombineMode(db->p++, G_CC_PRIMITIVE, G_CC_PRIMITIVE);
                 gSPTextureRectangle(db->p++,
@@ -716,7 +718,7 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
                 gDPSetCombineMode(db->p++, G_CC_MODULATEIA_PRIM, G_CC_MODULATEIA_PRIM);
             } else {
                 gDPSetPrimColor(db->p++, 0, 0, 0xFF, 0xFF, 0xFF, 0xFF);
-                text_print_size(db, d->name, left, top, font_width, font_height);
+                text_print_size(db, d->short_name, left, top, font_width, font_height);
             }
         }
 
@@ -729,7 +731,7 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
             sprite_load(db, &quest_items_sprite, 17, 1);
 
             for (int i = 0; i < rows; i++) {
-                dungeon_entry_t* d = &(dungeons[i]);
+                dungeon_entry_t* d = &(dungeons[i + (i > 11 ? 1 : 0)]); // skip Tower
                 if (!d->has_keys) continue;
 
                 int8_t current_keys = z64_file.dungeon_keys[d->index];
@@ -754,7 +756,7 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
             sprite_load(db, &quest_items_sprite, 14, 1);
 
             for (int i = 0; i < rows; i++) {
-                dungeon_entry_t* d = &(dungeons[i]);
+                dungeon_entry_t* d = &(dungeons[i + (i > 11 ? 1 : 0)]); // skip Tower
                 // Replace index 13 (Ganon's Castle) with 10 (Ganon's Tower)
                 int index = d->index == 13 ? 10 : d->index;
 
@@ -770,7 +772,7 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
             sprite_load(db, &quest_items_sprite, 10, 1);
 
             for (int i = 0; i < rows; i++) {
-                dungeon_entry_t* d = &(dungeons[i]);
+                dungeon_entry_t* d = &(dungeons[i + (i > 11 ? 1 : 0)]); // skip Tower
                 if (d->has_card && z64_file.gerudos_card) {
                     int top = start_top + ((icon_size + padding) * i);
                     sprite_draw(db, &quest_items_sprite, 0,
@@ -789,7 +791,7 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
             sprite_load(db, &quest_items_sprite, 16, 1);
 
             for (int i = 0; i < rows; i++) {
-                dungeon_entry_t* d = &(dungeons[i]);
+                dungeon_entry_t* d = &(dungeons[i + (i > 11 ? 1 : 0)]); // skip Tower
                 if (d->has_map && z64_file.dungeon_items[d->index].map) {
                     int top = start_top + ((icon_size + padding) * i);
                     sprite_draw(db, &quest_items_sprite, 0,
@@ -804,7 +806,7 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
             sprite_load(db, &quest_items_sprite, 15, 1);
 
             for (int i = 0; i < rows; i++) {
-                dungeon_entry_t* d = &(dungeons[i]);
+                dungeon_entry_t* d = &(dungeons[i + (i > 11 ? 1 : 0)]); // skip Tower
                 if (d->has_map && z64_file.dungeon_items[d->index].compass) {
                     int top = start_top + ((icon_size + padding) * i);
                     sprite_draw(db, &quest_items_sprite, 0,
@@ -821,7 +823,7 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
             sprite_load(db, &quest_items_sprite, 11, 1);
 
             for (int i = 0; i < rows; i++) {
-                dungeon_entry_t* d = &(dungeons[i]);
+                dungeon_entry_t* d = &(dungeons[i + (i > 11 ? 1 : 0)]); // skip Tower
                 if (d->skulltulas && z64_file.gs_flags[d->index ^ 0x03] == d->skulltulas) {
                     int top = start_top + ((icon_size + padding) * i);
                     sprite_draw(db, &quest_items_sprite, 0,
@@ -836,7 +838,7 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
 
         if (show_mq) {
             for (int i = 0; i < non_chest_game_dungeon_count; i++) {
-                dungeon_entry_t* d = &(dungeons[i]);
+                dungeon_entry_t* d = &(dungeons[i + (i > 11 ? 1 : 0)]); // skip Tower
                 if (CFG_DUNGEON_INFO_MQ_NEED_MAP && d->has_map &&
                         !z64_file.dungeon_items[d->index].map) {
                     continue;
@@ -855,7 +857,7 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
             sprite_load(db, &key_rupee_clock_sprite, 1, 1);
 
             for (int i = 0; i < dungeon_count; i++) {
-                dungeon_entry_t* d = &(dungeons[i]);
+                dungeon_entry_t* d = &(dungeons[i + (i > 11 ? 1 : 0)]); // skip Tower
                 bool show_silver_rupees = false;
                 uint8_t* silver_rupee_puzzles = CFG_DUNGEON_IS_MQ[d->index] ? d->silver_rupee_puzzles_mq : d->silver_rupee_puzzles_vanilla;
                 for (int puzzle_idx = 0; puzzle_idx < 4; puzzle_idx++) {
@@ -879,7 +881,7 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
             sprite_load(db, &font_sprite, 16, 10); // load characters 0 through 9
 
             for (int i = 0; i < dungeon_count; i++) {
-                dungeon_entry_t* d = &(dungeons[i]);
+                dungeon_entry_t* d = &(dungeons[i + (i > 11 ? 1 : 0)]); // skip Tower
                 bool show_silver_rupees = false;
                 uint8_t* silver_rupee_puzzles = CFG_DUNGEON_IS_MQ[d->index] ? d->silver_rupee_puzzles_mq : d->silver_rupee_puzzles_vanilla;
                 for (int puzzle_idx = 0; puzzle_idx < 4; puzzle_idx++) {
@@ -1177,9 +1179,9 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
         // Draw dungeon names
 
         for (int i = 0; i < rows; i++) {
-            dungeon_entry_t* d = &(dungeons[d_right_dungeon_idx(i)]); // skip Deku/DC/Jabu/Ice dynamically
+            dungeon_entry_t* d = &(dungeons[d_right_dungeon_idx(i)]); // skip Deku/DC/Jabu/Ice/Tower dynamically
             int top = start_top + ((icon_size + padding) * i) + 1;
-            text_print(db, d->name, left, top);
+            text_print(db, d->short_name, left, top);
         }
 
         left += ((SHUFFLE_CHEST_GAME == 1 ? 11 : 8) * font_sprite.tile_w) + padding;
@@ -1191,7 +1193,7 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
         sprite_load(db, &quest_items_sprite, 17, 1);
 
         for (int i = 0; i < rows; i++) {
-            dungeon_entry_t* d = &(dungeons[d_right_dungeon_idx(i)]); // skip Deku/DC/Jabu/Ice dynamically
+            dungeon_entry_t* d = &(dungeons[d_right_dungeon_idx(i)]); // skip Deku/DC/Jabu/Ice/Tower dynamically
             if (!d->has_keys) continue;
 
             int8_t current_keys = z64_file.dungeon_keys[d->index];
@@ -1216,7 +1218,7 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
         sprite_load(db, &quest_items_sprite, 14, 1);
 
         for (int i = 0; i < rows; i++) {
-            dungeon_entry_t* d = &(dungeons[d_right_dungeon_idx(i)]); // skip Deku/DC/Jabu/Ice dynamically
+            dungeon_entry_t* d = &(dungeons[d_right_dungeon_idx(i)]); // skip Deku/DC/Jabu/Ice/Tower dynamically
             // Replace index 13 (Ganon's Castle) with 10 (Ganon's Tower)
             int index = d->index == 13 ? 10 : d->index;
 
@@ -1232,7 +1234,7 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
         sprite_load(db, &quest_items_sprite, 10, 1);
 
         for (int i = 0; i < rows; i++) {
-            dungeon_entry_t* d = &(dungeons[d_right_dungeon_idx(i)]); // skip Deku/DC/Jabu/Ice dynamically
+            dungeon_entry_t* d = &(dungeons[d_right_dungeon_idx(i)]); // skip Deku/DC/Jabu/Ice/Tower dynamically
             if (d->has_card && z64_file.gerudos_card) {
                 int top = start_top + ((icon_size + padding) * i);
                 sprite_draw(db, &quest_items_sprite, 0,
@@ -1248,7 +1250,7 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
             sprite_load(db, &key_rupee_clock_sprite, 1, 1);
 
             for (int i = 0; i < rows; i++) {
-                dungeon_entry_t* d = &(dungeons[d_right_dungeon_idx(i)]); // skip Deku/DC/Jabu/Ice dynamically
+                dungeon_entry_t* d = &(dungeons[d_right_dungeon_idx(i)]); // skip Deku/DC/Jabu/Ice/Tower dynamically
                 bool show_silver_rupees = false;
                 uint8_t* silver_rupee_puzzles = CFG_DUNGEON_IS_MQ[d->index] ? d->silver_rupee_puzzles_mq : d->silver_rupee_puzzles_vanilla;
                 for (int puzzle_idx = 0; puzzle_idx < 4; puzzle_idx++) {
@@ -1272,7 +1274,7 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
             sprite_load(db, &font_sprite, 16, 10); // load characters 0 through 9
 
             for (int i = 0; i < rows; i++) {
-                dungeon_entry_t* d = &(dungeons[d_right_dungeon_idx(i)]);
+                dungeon_entry_t* d = &(dungeons[d_right_dungeon_idx(i)]); // skip Deku/DC/Jabu/Ice/Tower dynamically
                 bool show_silver_rupees = false;
                 uint8_t* silver_rupee_puzzles = CFG_DUNGEON_IS_MQ[d->index] ? d->silver_rupee_puzzles_mq : d->silver_rupee_puzzles_vanilla;
                 for (int puzzle_idx = 0; puzzle_idx < 4; puzzle_idx++) {
@@ -1339,13 +1341,21 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
 
         // Draw dungeon names
 
+        int d_idx = 0;
+
         for (int i = 0; i < 12; i++) {
-            dungeon_entry_t* d = &(dungeons[i + (i > 9 ? 1 : 0)]); // skip Hideout
+            // skip Hideout and Tower
+            if (d_idx == 10 || (d_idx == 12)) {
+                d_idx++;
+            }
+            dungeon_entry_t* d = &(dungeons[d_idx]);
+            d_idx++;
+
             bool empty = CFG_DUNGEON_PRECOMPLETED[d->index];
             int top = start_top + ((icon_size + padding) * i) + 1;
             if (empty) {
                 gDPSetPrimColor(db->p++, 0, 0, 0xFF, 0xFF, 0xFF, 0x7F);
-                uint16_t sizeRectangle = text_print(db, d->name, left, top) - left;
+                uint16_t sizeRectangle = text_print(db, d->short_name, left, top) - left;
                 gDPSetPrimColor(db->p++, 0, 0, 0xFF, 0xFF, 0xFF, 0xBF);
                 gDPSetCombineMode(db->p++, G_CC_PRIMITIVE, G_CC_PRIMITIVE);
                 gSPTextureRectangle(db->p++,
@@ -1357,7 +1367,7 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
                 gDPSetCombineMode(db->p++, G_CC_MODULATEIA_PRIM, G_CC_MODULATEIA_PRIM);
             } else {
                 gDPSetPrimColor(db->p++, 0, 0, 0xFF, 0xFF, 0xFF, 0xFF);
-                text_print(db, d->name, left, top);
+                text_print(db, d->short_name, left, top);
             }
         }
 
@@ -1371,8 +1381,15 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
 
             sprite_load(db, &quest_items_sprite, 16, 1);
 
+            d_idx = 0;
             for (int i = 0; i < 12; i++) {
-                dungeon_entry_t* d = &(dungeons[i + (i > 9 ? 1 : 0)]); // skip Hideout
+                // skip Hideout and Tower
+                if (d_idx == 10 || (d_idx == 12)) {
+                    d_idx++;
+                }
+                dungeon_entry_t* d = &(dungeons[d_idx]);
+                d_idx++;
+
                 if (d->has_map && z64_file.dungeon_items[d->index].map) {
                     int top = start_top + ((icon_size + padding) * i);
                     sprite_draw(db, &quest_items_sprite, 0,
@@ -1386,8 +1403,15 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
 
             sprite_load(db, &quest_items_sprite, 15, 1);
 
+            d_idx = 0;
             for (int i = 0; i < 12; i++) {
-                dungeon_entry_t* d = &(dungeons[i + (i > 9 ? 1 : 0)]); // skip Hideout
+                // skip Hideout and Tower
+                if (d_idx == 10 || (d_idx == 12)) {
+                    d_idx++;
+                }
+                dungeon_entry_t* d = &(dungeons[d_idx]);
+                d_idx++;
+
                 if (d->has_map && z64_file.dungeon_items[d->index].compass) {
                     int top = start_top + ((icon_size + padding) * i);
                     sprite_draw(db, &quest_items_sprite, 0,
@@ -1403,8 +1427,15 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
 
             sprite_load(db, &quest_items_sprite, 11, 1);
 
+            d_idx = 0;
             for (int i = 0; i < 12; i++) {
-                dungeon_entry_t* d = &(dungeons[i + (i > 9 ? 1 : 0)]); // skip Hideout
+                // skip Hideout and Tower
+                if (d_idx == 10 || (d_idx == 12)) {
+                    d_idx++;
+                }
+                dungeon_entry_t* d = &(dungeons[d_idx]);
+                d_idx++;
+
                 if (d->skulltulas && z64_file.gs_flags[d->index ^ 0x03] == d->skulltulas) {
                     int top = start_top + ((icon_size + padding) * i);
                     sprite_draw(db, &quest_items_sprite, 0,
@@ -1418,8 +1449,15 @@ void draw_dungeon_info(z64_disp_buf_t* db) {
         // Draw master quest dungeons
 
         if (show_mq) {
+            d_idx = 0;
             for (int i = 0; i < 12; i++) {
-                dungeon_entry_t* d = &(dungeons[i + (i > 9 ? 1 : 0)]); // skip Hideout
+                // skip Hideout and Tower
+                if (d_idx == 10 || (d_idx == 12)) {
+                    d_idx++;
+                }
+                dungeon_entry_t* d = &(dungeons[d_idx]);
+                d_idx++;
+
                 if (CFG_DUNGEON_INFO_MQ_NEED_MAP && d->has_map &&
                         !z64_file.dungeon_items[d->index].map) {
                     continue;