+++ Messages.py
@@ -1443,7 +1443,9 @@ def update_warp_song_text(messages: list[Message], world: World) -> None:
 
 def update_map_compass_messages(messages: list[Message], world: World):
     from Hints import HintArea, GossipText
-    if world.settings.enhance_map_compass and world.settings.shuffle_mapcompass != 'remove' and world.settings.world_count == 1:
+    maps_exist = world.settings.shuffle_map != 'remove'
+    compasses_exist = world.settings.shuffle_compass != 'remove'
+    if world.settings.enhance_map_compass and (maps_exist or compasses_exist) and world.settings.world_count == 1:
         dungeon_list = {
             #                      dungeon name                      compass map  name of the entrance leading to the boss
             'Deku Tree':          ("the \x05\x42Deku Tree",          0x62, 0x88, "Deku Tree Before Boss -> Queen Gohma Boss Room"),
@@ -1507,7 +1509,7 @@ def update_map_compass_messages(messages: list[Message], world: World):
         for dungeon in world.dungeons:
             if dungeon.name in ('Gerudo Training Ground', 'Ganons Castle'):
                 pass
-            elif dungeon.name in ('Bottom of the Well', 'Ice Cavern'):
+            elif dungeon.name in ('Bottom of the Well', 'Ice Cavern') and maps_exist:
                 dungeon_name, compass_id, map_id = dungeon_list[dungeon.name]
                 if 'map_dungeon_location' in world.settings.enhance_map_compass and world.settings.shuffle_dungeon_entrances != 'off':
                     if 'Dungeon' in world.mix_entrance_pools:
@@ -1534,35 +1536,46 @@ def update_map_compass_messages(messages: list[Message], world: World):
                             update_message_by_id(messages, map_id, map_message, allow_duplicates=True)
             else:
                 dungeon_name, compass_id, map_id, boss_entrance = dungeon_list[dungeon.name]
-                if 'compass_reward' in world.settings.enhance_map_compass:
-                    if world.settings.shuffle_dungeon_rewards != 'dungeon':
-                        if world.entrance_rando_reward_hints:
-                            vanilla_reward = world.get_location(dungeon.vanilla_boss_name).vanilla_item
-                            vanilla_reward_location = world.hinted_dungeon_reward_locations[vanilla_reward]
-                            if vanilla_reward_location is None:
-                                area = HintArea.ROOT
+                if compasses_exist:
+                    if 'compass_reward' in world.settings.enhance_map_compass:
+                        if world.settings.shuffle_dungeon_rewards != 'dungeon':
+                            if world.entrance_rando_reward_hints:
+                                vanilla_reward = world.get_location(dungeon.vanilla_boss_name).vanilla_item
+                                vanilla_reward_location = world.hinted_dungeon_reward_locations[vanilla_reward]
+                                if vanilla_reward_location is None:
+                                    area = HintArea.ROOT
+                                else:
+                                    area = HintArea.at(vanilla_reward_location)
+                                area = GossipText(area.text(world.settings.clearer_hints, preposition=True, use_2nd_person=True), [area.color], prefix='', capitalize=False)
+                                if 'compass_boss_location' in world.settings.enhance_map_compass and world.settings.shuffle_bosses != 'off':
+                                    if 'Boss' in world.mix_entrance_pools:
+                                        boss_area = HintArea.at(world.get_region(f'{dungeon.vanilla_boss_name} Boss Room'))
+                                        boss_area = GossipText(boss_area.text(world.settings.clearer_hints, preposition=True, use_2nd_person=True), [boss_area.color], prefix='', capitalize=False)
+                                        compass_message = str(GossipText(f"\x13\x75\x08You found the \x05\x41Compass\x05\x40 for \x01{dungeon_name}\x05\x40! {dungeon.vanilla_boss_name} is {boss_area}, and the {vanilla_reward} is {area}!\x09", prefix=''))
+                                    else:
+                                        boss_room = world.get_entrance(boss_entrance).connected_region.name
+                                        compass_message = f"\x13\x75\x08You found the \x05\x41Compass\x05\x40 for\x01{dungeon_name}\x05\x40! {boss_textboxes[boss_room]}\x05\x40\x01lurks, and the {vanilla_reward}\x01is {area}!\x09"
+                                else:
+                                    compass_message = f"\x13\x75\x08You found the \x05\x41Compass\x05\x40\x01for {dungeon_name}\x05\x40!\x01The {vanilla_reward} can be found\x01{area}!\x09"
                             else:
-                                area = HintArea.at(vanilla_reward_location)
-                            area = GossipText(area.text(world.settings.clearer_hints, preposition=True, use_2nd_person=True), [area.color], prefix='', capitalize=False)
+                                boss_location = next(filter(lambda loc: loc.type == 'Boss', world.get_entrance(f'{dungeon} Before Boss -> {dungeon.vanilla_boss_name} Boss Room').connected_region.locations))
+                                dungeon_reward = boss_location.item.name
+                                if 'compass_boss_location' in world.settings.enhance_map_compass and world.settings.shuffle_bosses != 'off':
+                                    boss_room = world.get_entrance(boss_entrance).connected_region.name
+                                    compass_message = f"\x13\x75\x08You found the \x05\x41Compass\x05\x40 for\x01{dungeon_name}\x05\x40!\x01In this dungeon, {boss_textboxes[boss_room]}\x05\x40\x01guards the \x05{COLOR_MAP[REWARD_COLORS[dungeon_reward]]}{dungeon_reward}\x05\x40!\x09"
+                                else:
+                                    compass_message = f"\x13\x75\x08You found the \x05\x41Compass\x05\x40\x01for {dungeon_name}\x05\x40!\x01It holds the \x05{COLOR_MAP[REWARD_COLORS[dungeon_reward]]}{dungeon_reward}\x05\x40!\x09"
+                            update_message_by_id(messages, compass_id, compass_message, allow_duplicates=True)
+                        else:
                             if 'compass_boss_location' in world.settings.enhance_map_compass and world.settings.shuffle_bosses != 'off':
                                 if 'Boss' in world.mix_entrance_pools:
                                     boss_area = HintArea.at(world.get_region(f'{dungeon.vanilla_boss_name} Boss Room'))
                                     boss_area = GossipText(boss_area.text(world.settings.clearer_hints, preposition=True, use_2nd_person=True), [boss_area.color], prefix='', capitalize=False)
-                                    compass_message = str(GossipText(f"\x13\x75\x08You found the \x05\x41Compass\x05\x40 for \x01{dungeon_name}\x05\x40! {dungeon.vanilla_boss_name} is {boss_area}, and the {vanilla_reward} is {area}!\x09", prefix=''))
+                                    compass_message = f"\x13\x75\x08You found the \x05\x41Compass\x05\x40 for\x01{dungeon_name}\x05\x40!\x01{dungeon.vanilla_boss_name} can be found\x01{boss_area}!\x09"
                                 else:
                                     boss_room = world.get_entrance(boss_entrance).connected_region.name
-                                    compass_message = f"\x13\x75\x08You found the \x05\x41Compass\x05\x40 for\x01{dungeon_name}\x05\x40! {boss_textboxes[boss_room]}\x05\x40\x01lurks, and the {vanilla_reward}\x01is {area}!\x09"
-                            else:
-                                compass_message = f"\x13\x75\x08You found the \x05\x41Compass\x05\x40\x01for {dungeon_name}\x05\x40!\x01The {vanilla_reward} can be found\x01{area}!\x09"
-                        else:
-                            boss_location = next(filter(lambda loc: loc.type == 'Boss', world.get_entrance(f'{dungeon} Before Boss -> {dungeon.vanilla_boss_name} Boss Room').connected_region.locations))
-                            dungeon_reward = boss_location.item.name
-                            if 'compass_boss_location' in world.settings.enhance_map_compass and world.settings.shuffle_bosses != 'off':
-                                boss_room = world.get_entrance(boss_entrance).connected_region.name
-                                compass_message = f"\x13\x75\x08You found the \x05\x41Compass\x05\x40 for\x01{dungeon_name}\x05\x40!\x01In this dungeon, {boss_textboxes[boss_room]}\x05\x40\x01guards the \x05{COLOR_MAP[REWARD_COLORS[dungeon_reward]]}{dungeon_reward}\x05\x40!\x09"
-                            else:
-                                compass_message = f"\x13\x75\x08You found the \x05\x41Compass\x05\x40\x01for {dungeon_name}\x05\x40!\x01It holds the \x05{COLOR_MAP[REWARD_COLORS[dungeon_reward]]}{dungeon_reward}\x05\x40!\x09"
-                        update_message_by_id(messages, compass_id, compass_message, allow_duplicates=True)
+                                    compass_message = f"\x13\x75\x08You found the \x05\x41Compass\x05\x40 for\x01{dungeon_name}\x05\x40! In this dungeon,\x01{boss_textboxes[boss_room]}\x05\x40 lurks!\x09"
+                                    update_message_by_id(messages, compass_id, compass_message, allow_duplicates=True)
                     else:
                         if 'compass_boss_location' in world.settings.enhance_map_compass and world.settings.shuffle_bosses != 'off':
                             if 'Boss' in world.mix_entrance_pools:
@@ -1573,34 +1586,25 @@ def update_map_compass_messages(messages: list[Message], world: World):
                                 boss_room = world.get_entrance(boss_entrance).connected_region.name
                                 compass_message = f"\x13\x75\x08You found the \x05\x41Compass\x05\x40 for\x01{dungeon_name}\x05\x40! In this dungeon,\x01{boss_textboxes[boss_room]}\x05\x40 lurks!\x09"
                                 update_message_by_id(messages, compass_id, compass_message, allow_duplicates=True)
-                else:
-                    if 'compass_boss_location' in world.settings.enhance_map_compass and world.settings.shuffle_bosses != 'off':
-                        if 'Boss' in world.mix_entrance_pools:
-                            boss_area = HintArea.at(world.get_region(f'{dungeon.vanilla_boss_name} Boss Room'))
-                            boss_area = GossipText(boss_area.text(world.settings.clearer_hints, preposition=True, use_2nd_person=True), [boss_area.color], prefix='', capitalize=False)
-                            compass_message = f"\x13\x75\x08You found the \x05\x41Compass\x05\x40 for\x01{dungeon_name}\x05\x40!\x01{dungeon.vanilla_boss_name} can be found\x01{boss_area}!\x09"
+                if maps_exist:
+                    if 'map_dungeon_location' in world.settings.enhance_map_compass and world.settings.shuffle_dungeon_entrances != 'off':
+                        dungeon_name = dungeon_name.removeprefix('the ') # to make room
+                        if 'Dungeon' in world.mix_entrance_pools:
+                            #TODO This won't work for Decoupled.
+                            area = HintArea.at(world.get_entrance(dungeon_entrances_reverse[dungeon.name]).connected_region)
+                            area = GossipText(area.text(world.settings.clearer_hints, preposition=True, use_2nd_person=True), [area.color], prefix='', capitalize=False)
+                            if 'map_mq' in world.settings.enhance_map_compass and (world.settings.mq_dungeons_mode == 'random' or world.settings.mq_dungeons_count != 0 and world.settings.mq_dungeons_count != 12):
+                                map_message = f"\x13\x76\x08You found the \x05\x41Map\x05\x40 for \x05{COLOR_MAP['Red'] + 'masterful' if world.dungeon_mq[dungeon.name] else COLOR_MAP['Green'] + 'ordinary'}\x05\x40\x01{dungeon_name}\x05\x40! This dungeon\x01is {area}\x05\x40!\x09"
+                            else:
+                                map_message = f"\x13\x76\x08You found the \x05\x41Map\x05\x40 for\x01{dungeon_name}\x05\x40! This dungeon is\x01{area}\x05\x40!\x09"
                         else:
-                            boss_room = world.get_entrance(boss_entrance).connected_region.name
-                            compass_message = f"\x13\x75\x08You found the \x05\x41Compass\x05\x40 for\x01{dungeon_name}\x05\x40! In this dungeon,\x01{boss_textboxes[boss_room]}\x05\x40 lurks!\x09"
-                            update_message_by_id(messages, compass_id, compass_message, allow_duplicates=True)
-                if 'map_dungeon_location' in world.settings.enhance_map_compass and world.settings.shuffle_dungeon_entrances != 'off':
-                    dungeon_name = dungeon_name.removeprefix('the ') # to make room
-                    if 'Dungeon' in world.mix_entrance_pools:
-                        #TODO This won't work for Decoupled.
-                        area = HintArea.at(world.get_entrance(dungeon_entrances_reverse[dungeon.name]).connected_region)
-                        area = GossipText(area.text(world.settings.clearer_hints, preposition=True, use_2nd_person=True), [area.color], prefix='', capitalize=False)
-                        if 'map_mq' in world.settings.enhance_map_compass and (world.settings.mq_dungeons_mode == 'random' or world.settings.mq_dungeons_count != 0 and world.settings.mq_dungeons_count != 12):
-                            map_message = f"\x13\x76\x08You found the \x05\x41Map\x05\x40 for \x05{COLOR_MAP['Red'] + 'masterful' if world.dungeon_mq[dungeon.name] else COLOR_MAP['Green'] + 'ordinary'}\x05\x40\x01{dungeon_name}\x05\x40! This dungeon\x01is {area}\x05\x40!\x09"
-                        else:
-                            map_message = f"\x13\x76\x08You found the \x05\x41Map\x05\x40 for\x01{dungeon_name}\x05\x40! This dungeon is\x01{area}\x05\x40!\x09"
-                    else:
-                        dungeon_index = [i for i, c in enumerate(dungeon_entrances) if dungeon.name in c]
-                        if 'map_mq' in world.settings.enhance_map_compass and (world.settings.mq_dungeons_mode == 'random' or world.settings.mq_dungeons_count != 0 and world.settings.mq_dungeons_count != 12):
-                            map_message = f"\x13\x76\x08You found the \x05\x41Map\x05\x40 for \x05{COLOR_MAP['Red'] + 'masterful' if world.dungeon_mq[dungeon.name] else COLOR_MAP['Green'] + 'ordinary'}\x05\x40\x01{dungeon_name}\x05\x40! This dungeon\x01is at {dungeon_textbox_list[dungeon_index[0]]}\x05\x40!\x09"
-                        else:
-                            map_message = f"\x13\x76\x08You found the \x05\x41Map\x05\x40 for\x01{dungeon_name}\x05\x40! This dungeon is\x01at {dungeon_textbox_list[dungeon_index[0]]}\x05\x40!\x09"
-                    update_message_by_id(messages, map_id, map_message, allow_duplicates=True)
-                else:
-                    if 'map_mq' in world.settings.enhance_map_compass and (world.settings.mq_dungeons_mode == 'random' or world.settings.mq_dungeons_count != 0 and world.settings.mq_dungeons_count != 12):
-                        map_message = f"\x13\x76\x08You found the \x05\x41Map\x05\x40 for\x01{dungeon_name}\x05\x40!\x01It\'s \x05{COLOR_MAP['Red'] + 'masterful' if world.dungeon_mq[dungeon.name] else COLOR_MAP['Green'] + 'ordinary'}\x05\x40!\x09"
+                            dungeon_index = [i for i, c in enumerate(dungeon_entrances) if dungeon.name in c]
+                            if 'map_mq' in world.settings.enhance_map_compass and (world.settings.mq_dungeons_mode == 'random' or world.settings.mq_dungeons_count != 0 and world.settings.mq_dungeons_count != 12):
+                                map_message = f"\x13\x76\x08You found the \x05\x41Map\x05\x40 for \x05{COLOR_MAP['Red'] + 'masterful' if world.dungeon_mq[dungeon.name] else COLOR_MAP['Green'] + 'ordinary'}\x05\x40\x01{dungeon_name}\x05\x40! This dungeon\x01is at {dungeon_textbox_list[dungeon_index[0]]}\x05\x40!\x09"
+                            else:
+                                map_message = f"\x13\x76\x08You found the \x05\x41Map\x05\x40 for\x01{dungeon_name}\x05\x40! This dungeon is\x01at {dungeon_textbox_list[dungeon_index[0]]}\x05\x40!\x09"
                         update_message_by_id(messages, map_id, map_message, allow_duplicates=True)
+                    else:
+                        if 'map_mq' in world.settings.enhance_map_compass and (world.settings.mq_dungeons_mode == 'random' or world.settings.mq_dungeons_count != 0 and world.settings.mq_dungeons_count != 12 and maps_exist):
+                            map_message = f"\x13\x76\x08You found the \x05\x41Map\x05\x40 for\x01{dungeon_name}\x05\x40!\x01It\'s \x05{COLOR_MAP['Red'] + 'masterful' if world.dungeon_mq[dungeon.name] else COLOR_MAP['Green'] + 'ordinary'}\x05\x40!\x09"
+                            update_message_by_id(messages, map_id, map_message, allow_duplicates=True)